feat: 리뷰 추가 및 삭제 기능 구현

+++ server/src/main/java/com/lifo/CVSreview/member/repository/MemberRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findById(int memberId);
+    Optional<Member> findById(long memberId);
 
 }
